+++ src/containerapp/HISTORY.rst
@@ -9,6 +9,7 @@ upcoming
 * 'az containerapp sessionpool create/update': Support `--lifecycle-type` and `--max-alive-period`
 * 'az containerapp up': support for `--kind` parameter
 * 'az containerapp env premium-ingress': Deprecate `--min-replicas` and `--max-replicas` parameters, use workload profile scale instead.
+* 'az containerapp sessionpool create/update': Support `--probe-yaml`
 
 1.2.0b3
 ++++++